+++ src/day4.rs
@@ -1,11 +1,11 @@
 use std::fs;
 
-pub fn solve() -> (u32, u32){
+pub fn solve() -> (u64, u64){
     let mut sol1 = 0;
     let mut sol2 = 0;
 
     let input = fs::read_to_string("inputs/4.txt").expect("File not found"); 
-    let mut dp:Vec<u32> = Vec::new();
+    let mut dp:Vec<u64> = Vec::new();
     for line in input.lines(){
         let (_, clean_line) = line.split_once(':').unwrap();
         let (wnumbs_str , ynumbs_str) = clean_line.split_once('|').unwrap();
@@ -31,7 +31,7 @@ pub fn solve() -> (u32, u32){
         dp.push(nr_matches);
         sol1 += end;
     }
-    let mut register:Vec<u32> = vec![1; dp.len()]; 
+    let mut register:Vec<u64> = vec![1; dp.len()]; 
     for i in 0..dp.len(){
         let j = dp[i];
         let r:usize = j.try_into().unwrap();
@@ -47,11 +47,11 @@ pub fn solve() -> (u32, u32){
     return (sol1, sol2);
 }
 
-fn parse_list(list:&str) -> Vec<u32>{
-    let mut ret:Vec<u32> = Vec::new(); 
+fn parse_list(list:&str) -> Vec<u64>{
+    let mut ret:Vec<u64> = Vec::new(); 
     let nr_list = list.split_whitespace();
     for nr_str in nr_list{
-        let val = nr_str.parse::<u32>().unwrap();
+        let val = nr_str.parse::<u64>().unwrap();
         ret.push(val);
     }
     return ret;
