+++ src/day6.rs
@@ -1,20 +1,20 @@
 use std::fs;
 
-pub fn solve() -> (u32, u32){
+pub fn solve() -> (u64, u64){
     let mut sol1 = 1;
     let input = fs::read_to_string("inputs/6.txt").unwrap();
     let mut lines = input.lines().into_iter();
     let (_, time_row_str) = lines.next().unwrap().split_once(':').unwrap();
     let (_, dist_row_str) = lines.next().unwrap().split_once(':').unwrap();
-    let time_row:Vec<u32> = time_row_str
+    let time_row:Vec<u64> = time_row_str
         .split_whitespace()
-        .map(|x| x.parse::<u32>())
+        .map(|x| x.parse::<u64>())
         .filter_map(Result::ok)
         .collect();
     
-    let dist_row:Vec<u32> = dist_row_str
+    let dist_row:Vec<u64> = dist_row_str
         .split_whitespace()
-        .map(|x| x.parse::<u32>())
+        .map(|x| x.parse::<u64>())
         .filter_map(Result::ok)
         .collect();
     
@@ -40,8 +40,8 @@ pub fn solve() -> (u32, u32){
     let mf1 = (sqrt - (t as f64))/(-2.0); 
     let mf2 = (-sqrt - (t as f64))/(-2.0);
     let solmf = (mf2 -mf1).floor();
-    let sol2 = solmf as u32;
-    // let mut win_count:u32 = 0;
+    let sol2 = solmf as u64;
+    // let mut win_count:u64 = 0;
     // for speed in 0..t{
     //     let run_time = t - speed;
     //     if speed * run_time > d{
